fix: eliminate remaining compiler warnings
- spatialGrid.h: add explicit int casts to fix C4244 double->int warnings
- simulation.cpp: cast [[nodiscard]] tick() return to void to fix C4834 warnings
- simulationRecorder.cpp: use gmtime_s instead of gmtime to fix C4996 deprecation
- seekerAgent.cpp: remove unreferenced catch variable to fix C4101 warning
- logger.h: use localtime_s/localtime_r instead of localtime to fix C4996 deprecation

+++ src/utils/spatialGrid.h
@@ -48,10 +48,10 @@ public:
      */
     std::vector<int> query(int row, int col, double radius) const {
         std::vector<int> result;
-        const int minR = cellCoord(row - radius);
-        const int maxR = cellCoord(row + radius);
-        const int minC = cellCoord(col - radius);
-        const int maxC = cellCoord(col + radius);
+        const int minR = cellCoord(static_cast<int>(row - radius));
+        const int maxR = cellCoord(static_cast<int>(row + radius));
+        const int minC = cellCoord(static_cast<int>(col - radius));
+        const int maxC = cellCoord(static_cast<int>(col + radius));
 
         for (int cr = minR; cr <= maxR; ++cr) {
             for (int cc = minC; cc <= maxC; ++cc) {
